pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pferdehof-sim"
version = "0.1.0"
description = "Cozy Discord horse adoption bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
